+++ components/ui/progress.tsx
@@ -12,7 +12,7 @@ const Progress = React.forwardRef<
   <ProgressPrimitive.Root
     ref={ref}
     className={cn(
-      "relative h-4 w-full overflow-hidden rounded-full bg-transparent border-black border-[1px]",
+      "relative h-4 w-full overflow-hidden rounded-full bg-transparent",
       className
     )}
     {...props}
